Login3 - UserActorFactory.test.ts
