+++ tests/UserActorFactory.test.ts
@@ -1,0 +1,17 @@
+import { UserActorFactory } from "@/lib/user/user-actor-factory";
+import { withDb } from "./_db_helpers";
+import { userPii, createUser, getUserActorFactoryConfig } from "./_fixtures";
+
+describe("UserActorFactory", () => {
+  describe("getUserActor", () => {
+    it("should load the UserActor for the corresponding email", async () => {
+      await withDb(async (db) => {
+        const user = await createUser(1, db);
+        const factory = new UserActorFactory(getUserActorFactoryConfig(db));
+        const pii = userPii(1);
+        const actor = await factory.getUserActor(pii.userEmail);
+        expect(actor?.getUserId()).toEqual(user.userId);
+      });
+    });
+  });
+});
